fix: linter issues
