+++ .pre-commit-config.yaml
@@ -47,6 +47,16 @@ repos:
         files: \.(ts|tsx)$
         pass_filenames: false
 
+  # Linting frontend
+  - repo: local
+    hooks:
+      - id: npm-run-lint
+        name: Npm Run Lint
+        entry: npm run lint -- --max-warnings=0
+        language: system
+        files: \.(ts|tsx)$
+        pass_filenames: false
+
   # Cargo fmt for Rust
   - repo: local
     hooks:
