+++ src/contexts/sentry-context.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+type SentryContextType = {
+  isInitialized: boolean;
+  error: string | null;
+};
+
+export const SentryContext = createContext<SentryContextType | undefined>(
+  undefined
+);
